background image and better styling with tailwindcss
