+++ src/components/Form/Label.tsx
@@ -3,7 +3,7 @@ import { LabelHTMLAttributes } from "react";
 export function Label(props: LabelHTMLAttributes<HTMLLabelElement>) {
   return (
     <label 
-      className="text-sm text-zinc-600 flex items-center justify-between"
+      className="text-sm text-white flex items-center justify-between"
       {...props}
     />
   )
