+++ src/components/FormHeader.tsx
@@ -1,0 +1,11 @@
+const FormHeader = () => (
+    <div className="flex flex-col h-24 items-center justify-center max-w-md bg-gray-600 w-full opacity-90">
+        <h1 className="mb-2 text-xl font-bold leading-tight tracking-tight text-gray-900 md:text-2xl dark:text-white">
+        Welcome to Company Register
+              </h1>
+        <p className=" mt-2 dark:text-white">Please fill up following details and get started with us.</p>
+    </div>
+  )
+  
+  export { FormHeader }
+
